Control.Concurrent.Async.Extra

Documentation

sequencePool :: Traversable t => Int -> t (IO a) -> IO (t a)

Implementation derived from Petr Pudlák's answer on StackOverflow http://stackoverflow.com/a/18898822/230050

mapPool :: Traversable t => Int -> (a -> IO b) -> t a -> IO (t b)

Implementation copied from Petr Pudlák's answer on StackOverflow http://stackoverflow.com/a/18898822/230050

sequenceConcurrently :: Traversable t => t (IO a) -> IO (t a)

forConcurrently_ :: Foldable t => t a -> (a -> IO b) -> IO ()

fixAsync :: (Async a -> IO a) -> IO (Async a)

Create an Async and pass it to itself.

fixAsyncBound :: (Async a -> IO a) -> IO (Async a)

Like fixAsync but using forkOS internally.

fixAsyncOn :: Int -> (Async a -> IO a) -> IO (Async a)

Like fixAsync but using forkOn internally.

fixAsyncWithUnmask :: (Async a -> (forall b. IO b -> IO b) -> IO a) -> IO (Async a)

Like fixAsync but using forkIOWithUnmask internally. The child thread is passed a function that can be used to unmask asynchronous exceptions.

fixAsyncOnWithUnmask :: Int -> (Async a -> (forall b. IO b -> IO b) -> IO a) -> IO (Async a)

Like fixAsyncOn but using forkOnWithUnmask internally. The child thread is passed a function that can be used to unmask asynchronous exceptions.

withParent :: Async a -> IO b -> IO (Async b)

Create an async that is linked to a parent. If the parent dies so does this async

newtype Promise a

Promise is like Concurrently but includes a sequential monad instance

Constructors

Promise

Fields

unPromise :: IO a